Refactor the broker validation
